Move actual Timer implementation to other file
This commit moves the native Timer implementation into another file. The goal is
to setup the ability to toggle between web-based and IO-based Timer
implementations.
